[GenSync] Introduce GenSync Interface
The GenSync interface is used as a framework to sync with distributed hosts. It supports continuous sync, and with multiple nodes. This commit contains a TCP connection interface, full sync, and IBLT sync protocol. It also updates the hash function and refactors the entire repo.

The TCP protocol is used to transfer data between nodes and the full sync is to sync data set between local and remote nodes. The IBLT is a probabilistic sync protocol to reconcile the set differences between two sets. The test also performs an IBLT success rate test which is dependent on the IBLT parameter setting.