IBLT Options
This commit introduces IBLT options including hashSync and datalen. The hash sync uses has function to sync the sets and exchange literal data depending on hashes. Without hashFunc, IBLT will exchange literal data. This commit also defult to using hash sync so that the knowledge of the max  datalen of the entire set is not needed.