[iblt] Allow multiple retries with larger tablesize
IBLT is a probablistic sync protocal. The size of the IBLT determins the probability of a successful sync. We allow retries on a failed sync attempt. The retries uses increased table size by 1 factor of their symetric difference. For example, to reconcile differences of 5, we might want to use rounded up 2.5 * diff number as the IBLT table size. The reties creates table of 3.5* diff number and 4.5, ..., etc, to increase the chance of successful sync upon every iteration. The IBLTs for retries are pre-computed and does not add up to sync time.